Refactor TextLineReader to core
