+++ swing/src/main/java/info/openrocket/swing/gui/help/tours/SlideSetManager.java
@@ -12,6 +12,7 @@ import java.util.Map;
 
 import javax.swing.text.html.StyleSheet;
 
+import info.openrocket.core.util.TextLineReader;
 import info.openrocket.swing.gui.util.GUIUtil;
 import info.openrocket.swing.gui.theme.UITheme;
